Socket Chat App
This is to explore the Socket.IO integration for DocWatch Repository
and as a small fun Chat application, that can run on most computers and
phones. This is the server side code. use npm to install all
dependencies and use node server.js to start it. Server runs on port
3000